+++ src/node/database/migrations/v1/create-bombs-exploded-table.ts
@@ -6,7 +6,7 @@ const createBombsExplodedTable: Migration = {
     await transaction.schema
       .createTable('bombs_exploded')
       .ifNotExists()
-      .addColumn('id', 'bigserial', (col) => col.primaryKey().notNull())
+      .addColumn('id', 'integer', (col) => col.primaryKey().notNull())
       .addColumn('match_checksum', 'varchar', (col) => col.notNull())
       .addForeignKeyConstraint('bombs_exploded_match_checksum_fk', ['match_checksum'], 'matches', ['checksum'], (cb) =>
         cb.onDelete('cascade'),
