+++ src/node/database/migrations/v1/create-damages-table.ts
@@ -6,7 +6,7 @@ const createDamagesTable: Migration = {
     await transaction.schema
       .createTable('damages')
       .ifNotExists()
-      .addColumn('id', 'bigserial', (col) => col.primaryKey().notNull())
+      .addColumn('id', 'integer', (col) => col.primaryKey().notNull())
       .addColumn('match_checksum', 'varchar', (col) => col.notNull())
       .addForeignKeyConstraint('damages_match_checksum_fk', ['match_checksum'], 'matches', ['checksum'], (cb) =>
         cb.onDelete('cascade'),
